+++ tests/testthat/test-aaa-deprecated.R
@@ -1,0 +1,8 @@
+context("deprecated")
+
+test_that("ssd_cfplot", {
+  setup(pdf(tempfile(fileext = ".pdf")))
+  teardown(dev.off())
+
+  expect_warning(ssd_cfplot(boron_data), "is deprecated as of")
+})