+++ NEWS.md
@@ -1,3 +1,4 @@
+- Deprecate `ssd_cfplot()` for `ssd_plot_cf()`.
 - Optimized `predict()` and added parallel argument.
 - Fix llog distribution with small concentrations.
 - Ensure concentrations below 1 have 1 significant figure in plots